profile & password

+++ database/migrations/2024_09_23_052741_create_bukus_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bukus', function (Blueprint $table) {
             $table->id();
             $table->string('judul');
-            $table->unsignedBigInteger('id_kategori');
+            $table->integer('id_kategori');
             $table->string('penulis');
             $table->integer('jml_hlmn');
             $table->string('penerbit');
